Debugging server commit ToInt convert

+++ main/management/commands/check_tour_ready.py
@@ -57,7 +57,7 @@ class Command(BaseCommand):
       print(is2)
       # Если являются степенью двойки то кол-во групп это половина от участников
       # Иначе находится ближайшее число степени двойки
-      groups = p_amount / 2 if is2 else self._find_closest_2rec(p_amount)
+      groups = int(p_amount / 2 if is2 else self._find_closest_2rec(p_amount))
       print(groups)
 
       return self._get_tour_table(particapants, groups, p_amount, is2)
